[Clarification] tries to clarify the play-money only nature of CPC.
Exceedingly few people have seen this document and nothing, links to
it.  Spending time to do even greater clarification is time wasted, so
I'll most likely remove this book from ctm.gihub.io, but I want to do
a deploy and then have archive.org archive this version before I
delete it.

It's not that I no longer think that Craft Poker Co.'s software will
eventually be used to deal real money poker on the internet, it's just
that doing so is so far removed from what we'll start out doing that
clarifying is too time consuming.